Some reorg and cleanup, affects more than just config code now
