+++ Adafruit_NeoPXL8.cpp
@@ -91,7 +91,7 @@
 // NeoPixel spec)...usually only affects the 1st pixel, subsequent pixels OK
 // due to signal reshaping through the 1st.
 
-static const int8_t defaultPins[] = {0, 1, 2, 3, 4, 5, 6, 7};
+static const int8_t defaultPins[] = NEOPXL8_DEFAULT_PINS;
 static volatile bool sending = 0;     // Set while DMA transfer is active
 static volatile uint32_t lastBitTime; // micros() when last bit issued
